feat(cli): multi-server config reading
