+++ packages/cli/src/commands/mcp-client.ts
@@ -1,4 +1,3 @@
-import { readFileSync } from "node:fs";
 import type { RawPrompt, RawResource, RawResourceTemplate, RawTool } from "@mcp-contracts/core";
 import { Client } from "@modelcontextprotocol/sdk/client/index.js";
 import { SSEClientTransport } from "@modelcontextprotocol/sdk/client/sse.js";
@@ -231,94 +230,4 @@ export async function captureServerData(client: Client): Promise<CapturedData> {
   }
 
   return { tools, resources, resourceTemplates, prompts };
-}
-
-/** Shape of a single server entry in mcp.json. */
-interface McpServerConfig {
-  command?: string;
-  args?: string[];
-  env?: Record<string, string>;
-  url?: string;
-}
-
-/**
- * Reads and resolves a server configuration from an mcp.json file.
- *
- * If only one server is defined and no --server is given, it's auto-selected.
- * If multiple servers exist, --server must be specified.
- *
- * @param configPath - Path to the mcp.json config file.
- * @param serverName - Optional server name to select.
- * @returns The resolved transport configuration.
- */
-export function readMcpConfig(
-  configPath: string,
-  serverName: string | undefined,
-): ResolvedTransport {
-  let raw: string;
-  try {
-    raw = readFileSync(configPath, "utf-8");
-  } catch (err) {
-    const message = err instanceof Error ? err.message : String(err);
-    throw new Error(`Failed to read config file "${configPath}": ${message}`);
-  }
-
-  let config: unknown;
-  try {
-    config = JSON.parse(raw);
-  } catch {
-    throw new Error(`Invalid JSON in config file "${configPath}"`);
-  }
-
-  if (typeof config !== "object" || config === null) {
-    throw new Error(`Config file "${configPath}" must contain a JSON object`);
-  }
-
-  const obj = config as Record<string, unknown>;
-  const servers = obj["mcpServers"] as Record<string, McpServerConfig> | undefined;
-
-  if (!servers || typeof servers !== "object") {
-    throw new Error(`Config file "${configPath}" is missing "mcpServers" object`);
-  }
-
-  const serverNames = Object.keys(servers);
-  if (serverNames.length === 0) {
-    throw new Error(`Config file "${configPath}" has no server entries`);
-  }
-
-  let selectedName: string;
-  if (serverName) {
-    if (!servers[serverName]) {
-      throw new Error(
-        `Server "${serverName}" not found in config. Available: ${serverNames.join(", ")}`,
-      );
-    }
-    selectedName = serverName;
-  } else if (serverNames.length === 1) {
-    selectedName = serverNames[0] as string;
-  } else {
-    throw new Error(
-      `Multiple servers in config. Use --server to select one: ${serverNames.join(", ")}`,
-    );
-  }
-
-  const entry = servers[selectedName];
-  if (!entry) {
-    throw new Error(`Server "${selectedName}" not found in config`);
-  }
-
-  if (entry.url) {
-    return { transport: "streamable-http", url: entry.url };
-  }
-
-  if (entry.command) {
-    return {
-      transport: "stdio",
-      command: entry.command,
-      args: entry.args,
-      env: entry.env,
-    };
-  }
-
-  throw new Error(`Server "${selectedName}" has neither "command" nor "url" configured`);
 }
